feat: 포인트 충전/잔액 기능 구현 및 단위 테스트 작성

+++ src/main/java/com/hanghae/concert_reservation/adapter/api/user/response/UserPointResponse.java
@@ -2,14 +2,16 @@ package com.hanghae.concert_reservation.adapter.api.user.response;
 
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class UserPointResponse {
 
     private Long userId;
 
-    private int balance;
+    private BigDecimal balance;
 
-    public UserPointResponse(Long userId, int balance) {
+    public UserPointResponse(Long userId, BigDecimal balance) {
         this.userId = userId;
         this.balance = balance;
     }
